Avoid code duplication in testing modules
Both printtest and proctest use methods to communicate with the underlying command.
These methods are common and shouldn't be implemented in both modules.
